+++ neoforge/src/main/java/io/github/mortuusars/envelope/neoforge/mixin/flammable/PigeonholeBlockMixin.java
@@ -1,5 +1,6 @@
 package io.github.mortuusars.envelope.neoforge.mixin.flammable;
 
+import io.github.mortuusars.envelope.Envelope;
 import io.github.mortuusars.envelope.world.block.PigeonholeBlock;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
@@ -13,11 +14,11 @@ import org.spongepowered.asm.mixin.Mixin;
 public class PigeonholeBlockMixin implements IBlockExtension {
     @Override
     public int getFlammability(@NotNull BlockState state, @NotNull BlockGetter level, @NotNull BlockPos pos, @NotNull Direction direction) {
-        return 20;
+        return state.is(Envelope.Tags.Blocks.PIGEONHOLES_THAT_BURN) ? 20 : 0;
     }
 
     @Override
     public int getFireSpreadSpeed(@NotNull BlockState state, @NotNull BlockGetter level, @NotNull BlockPos pos, @NotNull Direction direction) {
-        return 5;
+        return state.is(Envelope.Tags.Blocks.PIGEONHOLES_THAT_BURN) ? 5 : 0;
     }
 }